@@ -1,15 +1,15 @@
 package org.mackenzine.collections;
 
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class ListingTest {
 
     private Listing<Integer> listing;
 
-    @Before
+    @BeforeEach
     public void setUp() {
         listing = new Listing<>();
     }
